Pass indices as uint16 in tests

+++ models/demos/deepseek_v3_d_p/tests/pcc/test_prefill_dispatch.py
@@ -224,7 +224,11 @@ def test_ttnn_dispatch(
     )
 
     tt_indices = ttnn.from_torch(
-        indices, mesh_mapper=mesh_mapper_replicated, layout=ttnn.ROW_MAJOR_LAYOUT, device=mesh_device, dtype=ttnn.int32
+        indices.to(torch.int16),
+        mesh_mapper=mesh_mapper_replicated,
+        layout=ttnn.ROW_MAJOR_LAYOUT,
+        device=mesh_device,
+        dtype=ttnn.uint16,
     )
 
     # Create expert dispatch table
